requirements.txt
setuptools>=65.5.1
pyserial==3.5
customtkinter==5.2.0
pillow==10.0.0